Eliminate test duplication flagged by SonarQube in rubygems adapter tests
Extract shared error-handling tests and the some_gem_dependencies fixture into
SharedDependencyApiBehaviors module. Each adapter class includes the module and
implements stub_single_gem_request to satisfy the template method contract.
Unique test_get_list and adapter-specific setup remain in their respective files.
